riscv: added riscv profile to Flatcar
TBD: build an SDK and then build the packages.

Future work: decide on how the kernel / dtb / bootloader workflow.

Grub already has upstrea-ish support for RISCV64, but some of the
workflow needs to change to be able to boot on RISC-V physical boards.

The end goal though for this PR would be to be able to build a Flatcar
image that boots on QEMU-RISCV64.

Signed-off-by: Adrian Vladu <avladu@cloudbasesolutions.com>