github_actions: use minimal github actions to build Flatcar
The end goal is to have a minimal image build of Flatcar.
The time taken for such a build seems to be 3.5 hours.
The Flatcar image can be downloaded from the build artifacts tab (it is a zipped folder with the image).

To be able to build it, one need to enable github actions on their flatcar/scripts fork, add their code changes
and cherry-pick this commit on top of the changes.

Then, from the Actions tab of the repo, in this case: https://github.com/ader1990/scripts/actions,
select from the top left, Actions -> All workflows -> Run build -> Run workflow (from the top right) -> Select the desired branch -> Press `Run Workflow`.

Currently, the vendor format is qemu_uefi, but in the future, the code can be changed to produce other images.
Currently, the architecture can be chosen, default is arm64, and it supports amd64 too.